ai/requirements.txt
librosa>=0.10.0
numpy>=1.24.0
praat-parselmouth>=0.4.3
soundfile>=0.12.1
